backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch
torchvision
Pillow==10.4.0
numpy==2.1.3
python-multipart==0.0.9
tqdm==4.67.1
